# No pointer optimisation txn