--- a/mdbx/cursor.go
+++ b/mdbx/cursor.go
@@ -159,7 +159,7 @@ func (c *Cursor) Get(setkey, setval []byte, op uint) (key, val []byte, err error
 	}
 	if err != nil {
 		*c.txn.key = C.MDBX_val{}
-		*c.txn.val = C.MDBX_val{}
+		c.txn.val = C.MDBX_val{}
 		return nil, nil, err
 	}
 
@@ -174,12 +174,12 @@ func (c *Cursor) Get(setkey, setval []byte, op uint) (key, val []byte, err error
 			key = castToBytes(c.txn.key)
 		}
 	}
-	val = castToBytes(c.txn.val)
+	val = castToBytes(&c.txn.val)
 
 	// Clear transaction storage record storage area for future use and to
 	// prevent dangling references.
 	*c.txn.key = C.MDBX_val{}
-	*c.txn.val = C.MDBX_val{}
+	c.txn.val = C.MDBX_val{}
 
 	return key, val, nil
 }
@@ -189,7 +189,7 @@ func (c *Cursor) Get(setkey, setval []byte, op uint) (key, val []byte, err error
 //
 // See mdb_cursor_get.
 func (c *Cursor) getVal0(op uint) error {
-	ret := C.mdbx_cursor_get(c._c, c.txn.key, c.txn.val, C.MDBX_cursor_op(op))
+	ret := C.mdbx_cursor_get(c._c, c.txn.key, &c.txn.val, C.MDBX_cursor_op(op))
 	return operrno("mdbx_cursor_get", ret)
 }
 
@@ -206,7 +206,7 @@ func (c *Cursor) getVal1(setkey []byte, op uint) error {
 		c._c,
 		k, C.size_t(len(setkey)),
 		c.txn.key,
-		c.txn.val,
+		&c.txn.val,
 		C.MDBX_cursor_op(op),
 	)
 	return operrno("mdbx_cursor_get", ret)
@@ -220,7 +220,7 @@ func (c *Cursor) getVal01(setval []byte, op uint) error {
 		c._c,
 		v, C.size_t(len(setval)),
 		c.txn.key,
-		c.txn.val,
+		&c.txn.val,
 		C.MDBX_cursor_op(op),
 	)
 	return operrno("mdbx_cursor_get", ret)
@@ -242,7 +242,7 @@ func (c *Cursor) getVal2(setkey, setval []byte, op uint) error {
 		c._c,
 		k, C.size_t(len(setkey)),
 		v, C.size_t(len(setval)),
-		c.txn.key, c.txn.val,
+		c.txn.key, &c.txn.val,
 		C.MDBX_cursor_op(op),
 	)
 	return operrno("mdbx_cursor_get", ret)
@@ -280,16 +280,16 @@ func (c *Cursor) PutReserve(key []byte, n int, flags uint) ([]byte, error) {
 	ret := C.mdbxgo_cursor_put1(
 		c._c,
 		k, C.size_t(len(key)),
-		c.txn.val,
+		&c.txn.val,
 		C.MDBX_put_flags_t(flags|C.MDBX_RESERVE),
 	)
 	err := operrno("mdbx_cursor_put", ret)
 	if err != nil {
-		*c.txn.val = C.MDBX_val{}
+		c.txn.val = C.MDBX_val{}
 		return nil, err
 	}
-	b := castToBytes(c.txn.val)
-	*c.txn.val = C.MDBX_val{}
+	b := castToBytes(&c.txn.val)
+	c.txn.val = C.MDBX_val{}
 	return b, nil
 }