--- a/mdbx/txn_test.go
+++ b/mdbx/txn_test.go
@@ -1299,42 +1299,6 @@ func BenchmarkTxn_Get(b *testing.B) {
 	}
 }
 
-func BenchmarkTxn_GetFast(b *testing.B) {
-	env, _ := setup(b)
-
-	var db DBI
-	k := make([]byte, 8)
-	binary.BigEndian.PutUint64(k, uint64(1))
-
-	if err := env.Update(func(txn *Txn) (err error) {
-		db, err = txn.OpenRoot(0)
-		if err != nil {
-			return err
-		}
-		err = txn.Put(db, k, k, 0)
-		if err != nil {
-			return err
-		}
-		return nil
-	}); err != nil {
-		b.Errorf("dbi: %v", err)
-		return
-	}
-
-	if err := env.View(func(txn *Txn) (err error) {
-		b.ResetTimer()
-		for i := 0; i < b.N; i++ {
-			_, err := txn.GetFast(db, k)
-			if err != nil {
-				return err
-			}
-		}
-		return nil
-	}); err != nil {
-		b.Errorf("put: %v", err)
-	}
-}
-
 func TestTxnEnvWarmup(t *testing.T) {
 	env, _ := setup(t)