Added optimization that changed txn.val to non-pointer that allows us to avoid cgoCheckUnknownPointer false-positive cgo behavior. 
CI Check in Erigon: https://github.com/ledgerwatch/erigon/pull/10708