@JkLondon 2024-06-12
```
BenchmarkCursor-20                      10000000               163.9 ns/op            16 B/op          1 allocs/op
BenchmarkCursor_Renew/1-20              10000000                57.63 ns/op            0 B/op          0 allocs/op
BenchmarkCursor_Renew/2-20              10000000                55.57 ns/op            0 B/op          0 allocs/op
BenchmarkCursor_Renew/3-20              10000000               162.2 ns/op            16 B/op          1 allocs/op
BenchmarkCursor_Renew/4-20              10000000                60.87 ns/op            0 B/op          0 allocs/op
BenchmarkCursor_SetRange-20             10000000                97.85 ns/op            0 B/op          0 allocs/op
BenchmarkErrno_Error-20                 10000000               419.5 ns/op           224 B/op          4 allocs/op
BenchmarkTxn_abort-20                   10000000               220.2 ns/op            64 B/op          1 allocs/op
BenchmarkTxn_ro-20                      10000000               529.2 ns/op           208 B/op          4 allocs/op
BenchmarkTxn_unmanaged_abort-20         10000000               193.4 ns/op            64 B/op          1 allocs/op
BenchmarkTxn_unmanaged_commit-20        10000000               193.6 ns/op            64 B/op          1 allocs/op
BenchmarkTxn_unmanaged_ro-20            10000000               367.2 ns/op            96 B/op          3 allocs/op
BenchmarkTxn_renew-20                   10000000               118.9 ns/op             0 B/op          0 allocs/op
BenchmarkTxn_Put_append-20              --- FAIL: BenchmarkTxn_Put_append-20
    txn_test.go:1198: put: mdbx_put: MDBX_MAP_FULL: Environment mapsize limit reached
BenchmarkTxn_Put_append_noflag-20       --- FAIL: BenchmarkTxn_Put_append_noflag-20
    txn_test.go:1235: put: mdbx_put: MDBX_MAP_FULL: Environment mapsize limit reached
BenchmarkTxn_Get-20                     10000000                81.72 ns/op            0 B/op          0 allocs/op
```
master
```
BenchmarkCursor-20                      10000000               159.9 ns/op            16 B/op          1 allocs/op
BenchmarkCursor_Renew/1-20              10000000                55.68 ns/op            0 B/op          0 allocs/op
BenchmarkCursor_Renew/2-20              10000000                56.13 ns/op            0 B/op          0 allocs/op
BenchmarkCursor_Renew/3-20              10000000               167.3 ns/op            16 B/op          1 allocs/op
BenchmarkCursor_Renew/4-20              10000000                61.07 ns/op            0 B/op          0 allocs/op
BenchmarkCursor_SetRange-20             10000000                88.18 ns/op            0 B/op          0 allocs/op
BenchmarkErrno_Error-20                 10000000               419.3 ns/op           224 B/op          4 allocs/op
BenchmarkTxn_abort-20                   10000000               217.9 ns/op            64 B/op          1 allocs/op
BenchmarkTxn_ro-20                      10000000               546.7 ns/op           192 B/op          3 allocs/op
BenchmarkTxn_unmanaged_abort-20         10000000               198.7 ns/op            64 B/op          1 allocs/op
BenchmarkTxn_unmanaged_commit-20        10000000               199.0 ns/op            64 B/op          1 allocs/op
BenchmarkTxn_unmanaged_ro-20            10000000               334.1 ns/op            80 B/op          2 allocs/op
BenchmarkTxn_renew-20                   10000000               114.4 ns/op             0 B/op          0 allocs/op
BenchmarkTxn_Put_append-20              --- FAIL: BenchmarkTxn_Put_append-20
    txn_test.go:1198: put: mdbx_put: MDBX_MAP_FULL: Environment mapsize limit reached
BenchmarkTxn_Put_append_noflag-20       --- FAIL: BenchmarkTxn_Put_append_noflag-20
    txn_test.go:1235: put: mdbx_put: MDBX_MAP_FULL: Environment mapsize limit reached
BenchmarkTxn_Get-20                     10000000                71.83 ns/op            0 B/op          0 allocs/op
```
txn.val replacement (branch no-pointer-optimisation-txn)