@JkLondon 2024-06-12
```
goos: linux
goarch: amd64
pkg: github.com/erigontech/mdbx-go/mdbx
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
                        │ txn_val_pointer.txt │       txn_val_non_pointer.txt       │
                        │       sec/op        │   sec/op     vs base                │
Cursor-20                         168.2n ± 2%   166.6n ± 1%        ~ (p=0.289 n=10)
Cursor_Renew/1-20                 55.84n ± 2%   55.73n ± 1%        ~ (p=0.197 n=10)
Cursor_Renew/2-20                 55.86n ± 3%   55.20n ± 0%   -1.18% (p=0.003 n=10)
Cursor_Renew/3-20                 167.0n ± 1%   165.6n ± 2%        ~ (p=0.342 n=10)
Cursor_Renew/4-20                 61.15n ± 3%   62.22n ± 3%        ~ (p=0.529 n=10)
Cursor_SetRange-20                96.72n ± 3%   90.24n ± 2%   -6.70% (p=0.000 n=10)
Errno_Error-20                    383.7n ± 8%   377.3n ± 7%        ~ (p=0.436 n=10)
Txn_abort-20                      207.8n ± 1%   210.8n ± 2%        ~ (p=0.060 n=10)
Txn_ro-20                         465.2n ± 5%   463.2n ± 6%        ~ (p=0.353 n=10)
Txn_unmanaged_abort-20            192.8n ± 3%   195.5n ± 3%        ~ (p=0.210 n=10)
Txn_unmanaged_commit-20           191.2n ± 4%   198.2n ± 4%        ~ (p=0.468 n=10)
Txn_unmanaged_ro-20               311.1n ± 3%   278.1n ± 2%  -10.62% (p=0.000 n=10)
Txn_renew-20                      119.5n ± 5%   119.2n ± 6%        ~ (p=0.779 n=10)
Txn_Get-20                        83.56n ± 3%   73.20n ± 3%  -12.40% (p=0.000 n=10)
geomean                           147.0n        144.0n        -2.04%
```
compared benchmark of txn.val pointer (like in master) and non-pointer (this update)